backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "pydantic>=2.11.3",
    "typing>=3.10.0.0",
    "yt-dlp>=2025.3.31",
    "requests>=2.32.3",
    "python-dotenv>=1.1.0",
]
